Changed MemoryDiff to use polymorphic Change helper classes
There are still a couple of ugly dynamic_casts in the code to cope with
where there was a switch on type behaviour before, but largely the logic
is (or will be) and improvement.  There are also quite a lot of
dynamic_casts in the test code, but these may never change as it is the
sort of thing that test code needs to do.