CreateProcess dialog prepends command to the args string
Most process expect the executable name to be argv[0]. To match the
Visual Studio debug arguments convention we don't make users type this
in manually. This does means we now don't have a way for users to
specify a custom argv[0], but this is unlikely to ever be an issue.